requirements.txt
beautifulsoup4>=4.4.1
bs4>=0.0.1
future==0.15.2
lxml==3.5.0
python-telegram-bot>=9.0
requests==2.9.1
